Refactor page content types
