+++ components/PhpPageContainer.php
@@ -2,13 +2,13 @@
 
 namespace humhub\modules\custom_pages\components;
 
-use HttpException;
 use humhub\modules\custom_pages\models\forms\SettingsForm;
 use humhub\modules\custom_pages\models\CustomPage;
-use humhub\modules\custom_pages\models\PhpType;
+use humhub\modules\custom_pages\types\PhpType;
 use humhub\modules\file\libs\FileHelper;
 use Yii;
 use yii\helpers\Html;
+use yii\web\ForbiddenHttpException;
 
 /**
  * @used-by CustomPage
@@ -27,7 +27,7 @@ trait PhpPageContainer
         if (PhpType::isType($this->type)) {
             $settings = new SettingsForm();
             if ($this->isNewRecord && !$settings->phpPagesActive) {
-                throw new HttpException(403);
+                throw new ForbiddenHttpException();
             }
 
             if (!$this->validatePhpViewFile()) {
