+++ models/CustomPage.php
@@ -24,6 +24,13 @@ use humhub\modules\custom_pages\modules\template\components\TemplateRenderer;
 use humhub\modules\custom_pages\modules\template\models\Template;
 use humhub\modules\custom_pages\modules\template\models\TemplateInstance;
 use humhub\modules\custom_pages\permissions\ManagePages;
+use humhub\modules\custom_pages\types\ContentType;
+use humhub\modules\custom_pages\types\HtmlType;
+use humhub\modules\custom_pages\types\IframeType;
+use humhub\modules\custom_pages\types\LinkType;
+use humhub\modules\custom_pages\types\MarkdownType;
+use humhub\modules\custom_pages\types\PhpType;
+use humhub\modules\custom_pages\types\TemplateType;
 use humhub\modules\custom_pages\widgets\WallEntry;
 use humhub\modules\space\models\Space;
 use humhub\modules\user\helpers\AuthHelper;
