+++ types/TemplateType.php
@@ -6,69 +6,31 @@
  * Time: 13:29
  */
 
-namespace humhub\modules\custom_pages\models;
+namespace humhub\modules\custom_pages\types;
 
+use humhub\modules\custom_pages\models\CustomPage;
 use humhub\modules\custom_pages\modules\template\components\TemplateCache;
 use humhub\modules\custom_pages\modules\template\models\TemplateInstance;
 use humhub\modules\custom_pages\modules\template\models\PagePermission;
-use yii\widgets\ActiveForm;
 use Yii;
+use yii\widgets\ActiveForm;
 use yii\base\InvalidArgumentException;
 
 class TemplateType extends ContentType
 {
     public const ID = 5;
 
-    public function getId()
-    {
-        return static::ID;
-    }
-
-    public function getLabel()
+    public function getLabel(): string
     {
         return Yii::t('CustomPagesModule.base', 'Template');
     }
 
-    public function getDescription()
+    public function getDescription(): string
     {
         return Yii::t('CustomPagesModule.base', 'Templates allow you to define combinable page fragments with inline edit functionality.');
     }
 
-    /**
-     * @param CustomPage $page
-     * @param bool $insert
-     * @param array $changedAttributes
-     * @return bool
-     */
-    public function afterSave($page, $insert, $changedAttributes)
-    {
-        if (!parent::afterSave($page, $insert, $changedAttributes)) {
-            return false;
-        }
-
-        if ($insert) {
-            $templateInstance = new TemplateInstance([
-                'page_id' => $page->id,
-                'template_id' => $page->templateId,
-            ]);
-            return $templateInstance->save();
-        }
-
-        return true;
-
-    }
-
-    public function afterDelete($page)
-    {
-        TemplateInstance::deleteByOwner($page);
-    }
-
-    /**
-     * @param CustomPage $content
-     * @param array $options
-     * @return string
-     */
-    public function render(CustomPage $content, $options = [])
+    public function render(CustomPage $content, $options = []): string
     {
         $templateInstance = TemplateInstance::findOne(['page_id' => $content->id]);
 
@@ -92,13 +54,38 @@ class TemplateType extends ContentType
         return $html;
     }
 
-    public function getViewName()
+    public function getViewName(): string
     {
         return 'template';
     }
 
-    public function renderFormField(ActiveForm $form, CustomPage $page)
+    public function renderFormField(ActiveForm $form, CustomPage $page): string
     {
         return $form->field($page, 'templateId')->dropDownList($page->getAllowedTemplateSelection(), ['value' => $page->getTemplateId(), 'disabled' => !$page->isNewRecord]);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function afterSave(CustomPage $page, bool $insert, array $changedAttributes): bool
+    {
+        if (!parent::afterSave($page, $insert, $changedAttributes)) {
+            return false;
+        }
+
+        if ($insert) {
+            $templateInstance = new TemplateInstance([
+                'page_id' => $page->id,
+                'template_id' => $page->templateId,
+            ]);
+            return $templateInstance->save();
+        }
+
+        return true;
+    }
+
+    public function afterDelete(CustomPage $page): void
+    {
+        TemplateInstance::deleteByOwner($page);
     }
 }
