+++ views/common/add.php
@@ -1,7 +1,7 @@
 <?php
 
 use humhub\modules\custom_pages\helpers\Url;
-use humhub\modules\custom_pages\models\ContentType;
+use humhub\modules\custom_pages\types\ContentType;
 use humhub\modules\custom_pages\widgets\AddContentTypeRow;
 use humhub\widgets\Button;
