+++ widgets/views/addContentTypeRow.php
@@ -6,7 +6,7 @@ use yii\helpers\Html;
 use humhub\modules\content\helpers\ContentContainerHelper;
 
 /** @var \humhub\modules\custom_pages\models\Target $target */
-/** @var \humhub\modules\custom_pages\models\ContentType $contentType*/
+/** @var \humhub\modules\custom_pages\types\ContentType $contentType*/
 /** @var string $pageType */
 /** @var bool $disabled*/
